pyproject.toml
[project]
name = "jasp-multimodal-rag"
version = "0.1.0"
description = "Multimodal RAG backend for the JASP manual assistant"
authors = [
  { name = "Yuan, Vivi (BDS) – UvA", email = "ywxiu0811@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.13"

dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "streamlit",
  "pydantic",
  "pyyaml",
  "python-dotenv",
  "langchain",
  "langchain-community",
  "chromadb",
  "sentence-transformers",
  "transformers",
  "open-clip-torch",
  "pillow",
  "pymupdf",
  "tqdm",
  "pypdf (>=6.0.0,<7.0.0)",
  "youtube-transcript-api (>=0.6.2,<0.7.0)",
  "pytube (>=15.0.0,<16.0.0)",
  "yt-dlp (>=2025.9.23,<2026.0.0)",
  "openai-whisper @ git+https://github.com/openai/whisper.git",
  "watchdog (>=6.0.0,<7.0.0)"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "black",
  "ruff"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false


